Ajout d'une gestion d'erreur complète Gestion des accès au page administrateur seulement Correctif sur l'outil getliste afin de pouvoir faire des conditions like ou =

+++ pages/loginRegister/Remove.php
@@ -1,9 +1,13 @@
 <?php
-
+session_start();
 require '../Fonctions.php';
 $bdd = getDataBase();
-$query = "DELETE FROM membres WHERE id=". $_POST["id"];
-$statement = $bdd->prepare($query);
-$statement->execute();
-$statement->closeCursor();
+if (isset($bdd)){
+    $query = "DELETE FROM membres WHERE id=". $_POST["id"];
+    $statement = $bdd->prepare($query);
+    $statement->execute();
+    $statement->closeCursor();
+} else {
+    $_SESSION["erreur"]=7;
+}
 header('Location: ../GérerMembres.php');
